rename main Evm structs, introduce wip external type
