+++ Cargo.toml
@@ -1,6 +1,13 @@
 [workspace]
+members = [
+    #"bins/revme",
+    "bins/revm-test",
+    "crates/revm",
+    "crates/primitives",
+    "crates/interpreter",
+    "crates/precompile",
+]
 resolver = "2"
-members = ["bins/*", "crates/*"]
 default-members = ["crates/revm"]
 
 [profile.release]
