+++ crates/revm/src/frame.rs
@@ -1,3 +1,5 @@
+use revm_interpreter::InterpreterResult;
+
 use crate::JournalCheckpoint;
 use crate::{interpreter::Interpreter, primitives::Address};
 use core::ops::Range;
@@ -16,4 +18,20 @@ pub struct CallStackFrame {
     pub subcall_return_memory_range: Range<usize>,
     /// Interpreter
     pub interpreter: Interpreter,
+}
+
+
+/// Contains either a frame or a result.
+pub enum FrameOrResult {
+    /// Boxed stack frame
+    Frame(Box<CallStackFrame>),
+    /// Interpreter result
+    Result(InterpreterResult),
+}
+
+impl FrameOrResult {
+    /// Returns new frame.
+    pub fn new_frame(frame: CallStackFrame) -> Self {
+        Self::Frame(Box::new(frame))
+    }
 }
