+++ crates/revm/src/inspector/instruction.rs
@@ -1,4 +1,4 @@
-use crate::EVMImpl;
+use crate::Evm;
 use alloc::boxed::Box;
 use revm_interpreter::{
     opcode::{BoxedInstruction, Instruction},
@@ -7,11 +7,11 @@ use revm_interpreter::{
 };
 
 /// Outer closure that calls Inspector for every instruction.
-pub fn inspector_instruction<'a, SPEC: Spec + 'static, DB: Database>(
-    instruction: Instruction<EVMImpl<'a, SPEC, DB>>,
-) -> BoxedInstruction<'a, EVMImpl<'a, SPEC, DB>> {
+pub fn inspector_instruction<'a, SPEC: Spec + 'static, EXT, DB: Database>(
+    instruction: Instruction<Evm<'a, SPEC, EXT, DB>>,
+) -> BoxedInstruction<'a, Evm<'a, SPEC, EXT, DB>> {
     Box::new(
-        move |interpreter: &mut Interpreter, host: &mut EVMImpl<'a, SPEC, DB>| {
+        move |interpreter: &mut Interpreter, host: &mut Evm<'a, SPEC, EXT, DB>| {
             if let Some(inspector) = host.inspector.as_mut() {
                 // SAFETY: as the PC was already incremented we need to subtract 1 to preserve the
                 // old Inspector behavior.
@@ -39,15 +39,15 @@ pub fn inspector_instruction<'a, SPEC: Spec + 'static, DB: Database>(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{db::EmptyDB, interpreter::opcode::*, primitives::BerlinSpec, EVMImpl};
+    use crate::{db::EmptyDB, interpreter::opcode::*, primitives::BerlinSpec, Evm};
 
     #[test]
     fn test_make_boxed_instruction_table() {
         // test that this pattern builds.
-        let inst: InstructionTable<EVMImpl<'_, BerlinSpec, EmptyDB>> =
-            make_instruction_table::<EVMImpl<'_, BerlinSpec, _>, BerlinSpec>();
-        let _test: BoxedInstructionTable<'_, EVMImpl<'_, BerlinSpec, _>> =
-            make_boxed_instruction_table::<'_, EVMImpl<'_, BerlinSpec, EmptyDB>, BerlinSpec, _>(
+        let inst: InstructionTable<Evm<'_, BerlinSpec, (), EmptyDB>> =
+            make_instruction_table::<Evm<'_, BerlinSpec, _, _>, BerlinSpec>();
+        let _test: BoxedInstructionTable<'_, Evm<'_, BerlinSpec, _, _>> =
+            make_boxed_instruction_table::<'_, Evm<'_, BerlinSpec, _, EmptyDB>, BerlinSpec, _>(
                 inst,
                 inspector_instruction,
             );
